Added controller_namesapce to the left_wheel_names and right_wheel_names parameters
The controller namespace is added to the odom_frame_id and the base_frame_id. However, it is not done to the left_wheel_names and right_wheel_names.

The differential drive controller is often used as a gazebo plugin like below. Since this is defined inside a xacro, the Rewritten yaml cannot be used to add namespaces to joints. Adding the namespace in the cpp file bypasses this issue. 
In the PR, the convention is followed. That is, "robot_namespace/left_wheel_name".
Please let me know if I need to change it in any other place. The modified changes solves my issue.
```
    <gazebo>
      <plugin filename="libgazebo_ros2_control.so" name="gazebo_ros2_control">
        <namespace>${robot_namespace}</namespace>
        <robot_param>robot_description</robot_param>
        <robot_param_node>robot_state_publisher</robot_param_node>
        <parameters>package://scout_description/config/robot_control.yaml</parameters>
      </plugin>
    </gazebo>
```