Check goal tolerance from error to next command
We calculate the state error at the beginning of the control cycle
based on the estimated progress in the trajectory (respecting scaling)
compared to the current joint state.

When we are approaching the final point, that state error might be 0,
though we haven't reached our goal, yet. In this case the trajectory
would succeed and falsely report success in the case of a strict goal
tolerance.

This change checks the current state against the command once
interpolation reached the last point in the trajectory. This allows the
robot to arrive at the final point as long as it is doing so within the
configured goal time limit.