Fill point_before_trajectory with same information as trajectory
When executing a trajectory with a start time in the future, the
trajectory interpolation will create a setpoint at the robot's current
configuration based on the state interface information.

However, when the state interface for example doesn't have acceleration
information, but the trajectory does contain position, velocity and
acceleration, the segment from the current state to the first trajectory
point will not use quintic spline interpolation, though the rest of the
trajectory will.

This commit fills the current state with velocity and acceleration
information if it is not given in the state interfaces but it is defined
in the trajectory's first point.